replaced `long long` with generic `hiop::int_type` (#242)
* replaced 'long long' with generic hiop::index_type and hiop::size_type 

* C interface uses index and size int types

* moved hiop types to a header shared by both C and C++

* MPI Datatype was set to generic HiOp index and size integer types

* formatting with C++ stream to take advantage of overloading

+++ tests/LinAlg/testBase.hpp
@@ -65,7 +65,7 @@ namespace hiop { namespace tests {
 
 using real_type             = double;
 using local_ordinal_type    = int;
-using global_ordinal_type   = long long;
+using global_ordinal_type   = int;
 
 static const real_type zero = 0.0;
 static const real_type quarter = 0.25;
